+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Improved Range string conversion
+
 ## v0.1.1 (2022-09-09)
 * Added isV* methods to Scopes